Implemented dispy for job distribution; Implemented sqlite database for saving results; Implemented dispy for sending sweep spaces

+++ src/quartus/setup.py
@@ -5,7 +5,7 @@ import time
 
 class ProgramTunerWrapper(MeasurementInterface):
 
-  def get_qor(self):
+  def get_qor(self, sweepparam):
     f = open('sta.summary', 'r')
     while True:
       line = f.readline()
@@ -37,7 +37,7 @@ class ProgramTunerWrapper(MeasurementInterface):
 
     metadata = [comb_alut, mem_alut, reg, bram, dsp]
     return -float(slack), metadata
-    #return 0.123, ['1', '2', '3', '4', '5']
+    #return 0.123, [sweepparam, '2', '3', '4', '5']
 
   def run(self, desired_result, input, limit):
     """
@@ -62,7 +62,6 @@ class ProgramTunerWrapper(MeasurementInterface):
     #]
 
     cfg = desired_result.configuration.data
-    print str(cfg)
     result_id = desired_result.id
 
     f = open('./options.tcl', 'w')
@@ -76,24 +75,34 @@ class ProgramTunerWrapper(MeasurementInterface):
     f.write('"\n')
     f.close()
 
-    # delete previous results
-    cleanupcmd = 'rm localresult.txt'
-    subprocess.Popen(cleanupcmd, shell=True).wait()
+    if 'sweepparam' in cfg:
+        # delete previous results
+        cleanupcmd = 'rm localresult.txt'
+        subprocess.Popen(cleanupcmd, shell=True).wait()
 
-    print "Starting " + cfg['sweepparam']
-    print "Changing param"
-    tclmodcmd = 'sed \'s/SWEEPPARAM/BITS ' + cfg['sweepparam'] + '/g\' run_quartus.tcl > run_quartus_sweep.tcl'
-    subprocess.Popen(tclmodcmd, shell=True).wait()
-    print "Starting compilation"
-    cmd = 'quartus_sh -t ./run_quartus_sweep.tcl'
-    run_result = self.call_program(cmd)
-    assert run_result['returncode'] == 0
-    print "Finished compilation"
-    result, metadata = self.get_qor()
-    self.dumpresult(cfg, result, metadata)
-    print "Finished " + cfg['sweepparam']
-    cleanupcmd = 'rm run_quartus_sweep.tcl'
-    subprocess.Popen(cleanupcmd, shell=True).wait()
-    print "Clean up done"
+        # generate verilog design file; this is to integrate the libcharm genverilog scripts
+        sweepparam = int(cfg['sweepparam'])
+        genveri = 'cd design; python genVerilogMultiplier.py ' + str(sweepparam) + ' ' + str(sweepparam + 1) + ' 1; cd ..'
+        subprocess.Popen(genveri, shell=True).wait()
+
+        print "Starting " + cfg['sweepparam']
+        tclmodcmd = 'sed \'s/SWEEPPARAM/BITS ' + cfg['sweepparam'] + '/g\' run_quartus.tcl > run_quartus_sweep.tcl'
+        subprocess.Popen(tclmodcmd, shell=True).wait()
+        cmd = 'quartus_sh -t ./run_quartus_sweep.tcl'
+        #cmd = 'ls'
+        run_result = self.call_program(cmd)
+        assert run_result['returncode'] == 0
+        result, metadata = self.get_qor(cfg['sweepparam'])
+        self.dumpresult(cfg, cfg['sweepparam'], result, metadata)
+        cleanupcmd = 'rm run_quartus_sweep.tcl'
+        subprocess.Popen(cleanupcmd, shell=True).wait()
+        print "Finished " + cfg['sweepparam']
+    else:
+        cmd = 'quartus_sh -t ./run_quartus.tcl'
+        run_result = self.call_program(cmd)
+        assert run_result['returncode'] == 0
+
+        result, metadata = self.get_qor()
+        self.dumpresult(cfg, result, metadata)
 
     return Result(time = result)
